+++ buttons/configuration.go
@@ -1,0 +1,10 @@
+package buttons
+
+type Configuration struct {
+	Switches map[string]ButtonConfiguration `json:"switches"`
+}
+
+type ButtonConfiguration struct {
+	Gpio string `json:"gpio"`
+	name string
+}