+++ entrypoint/entrypoint.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"log"
 
+	"github.com/kilianp07/AthleteIQBox/buttons"
 	"github.com/kilianp07/AthleteIQBox/gps"
 	"github.com/kilianp07/AthleteIQBox/transmitter"
 	utils "github.com/kilianp07/AthleteIQBox/utils/logger"
@@ -16,6 +17,7 @@ type configuration struct {
 	Gps       gps.Configuration         `json:"gps"`
 	Bluetooth transmitter.Conf          `json:"bluetooth"`
 	Logger    utils.LoggerConfiguration `json:"logger"`
+	Buttons   buttons.Configuration     `json:"buttons"`
 }
 
 func Start(confFile string) {
@@ -39,13 +41,15 @@ func Start(confFile string) {
 	// Configure logger.
 	logger = utils.NewLogger(conf.Logger)
 
-	if err = gps.Configure(conf.Gps); err != nil {
-		logger.Errorf("error configuring gps reader: %v", err)
+	// Configure buttons
+	buttonsManager, err := buttons.NewManager(conf.Buttons)
+	if err != nil {
+		logger.Errorf("error creating buttons manager: %v", err)
 		return
 	}
 
-	if err := gps.Start(); err != nil {
-		logger.Errorf("error starting gps reader: %v", err)
+	if err = gps.Configure(conf.Gps); err != nil {
+		logger.Errorf("error configuring gps reader: %v", err)
 		return
 	}
 
@@ -55,11 +59,33 @@ func Start(confFile string) {
 		return
 	}
 
-	if err = t.Start(context.Background()); err != nil {
-		logger.Errorf("error starting transmitter: %v", err)
+	if err := t.Start(context.Background()); err != nil {
+		logger.Errorf("Error starting transmitter: %v", err)
 		return
 	}
 
-	// Block forever.
-	select {}
+	// Block and wait for button presses indefinitely
+	logger.Debugf("Entering button press loop")
+	for r := range buttonsManager.Run() {
+		logger.Infof("Received button press event: %v", r)
+
+		if r {
+			logger.Debugf("Handling start button press")
+			logger.Infof("Transmitter started successfully")
+
+			if err := gps.Start(); err != nil {
+				logger.Errorf("Error starting GPS reader: %v", err)
+				return
+			}
+			logger.Infof("GPS reader started successfully")
+		} else {
+			logger.Infof("Handling stop button press")
+			if err := gps.Stop(); err != nil {
+				logger.Errorf("Error stopping GPS reader: %v", err)
+				return
+			}
+			logger.Infof("GPS reader stopped successfully")
+		}
+	}
+
 }